Compute Krippendorff's Replicability (formerly Krippendorff Alpha)

This function implements the computation of Krippendorff's Alpha as per https://repository.upenn.edu/cgi/viewcontent.cgi?article=1043&context=asc_papers # nolint

replicability(dt, unit, measurement, level)

Arguments

dt
`data.table` containing the reliability data in long format.

unit
Name of the column containing the unit ID

measurement
Name of the column containing the measurements, one per judge

level
c('binary', 'nominal'): type of oberverment data.

Value

alpha

Krippendorff's Alpha reliability index

De

Expected disagreement

Do

Overall observed disagreement accross all units

by_unit

Dataframe with one line per unit and columns

unit

Unit

mu

Number of observations in that unit

Do

Observed disagreement within this unit

Details

It is designed to be space efficient for sparse oberverments, and as thus does not take as input a reliability matrix, but a long-format data.table

If a tibble or a non-data.table dataframe is passed as input, this function will still work! It will silently create a data.table copy of your dataframe (or tibble). TODO(jucor): cite properly